Moved existing code to the framework package
