+++ wolf_engine_framework/src/logging.rs
