+++ wolf_engine_framework/src/plugins/core_plugin.rs
