+++ wolf_engine_framework/src/schedulers/simple_render_scheduler.rs
